requirements.txt
aiogram>=3.7.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic>=2.5.2
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
requests>=2.31.0
gspread>=5.12.4
google-auth>=2.23.4
alembic>=1.13.0
rapidfuzz==3.10.0
aiohttp>=3.9.5
asyncpg>=0.29.0
pymorphy3==1.2.0
pymorphy3-dicts-uk==2.4.1.1.1663094765